Tests: Moving test3d to tests directory and update central opesci code
