+++ include/opesciIO.h
@@ -48,6 +48,7 @@ void opesci_dump_receivers_vts(std::string name, const int dims[], const float s
 			       std::vector<float> &uss, std::vector<float> &vss, std::vector<float> &wss, std::vector<float> &pss);
 
 int opesci_read_simple_binary(const char *filename, std::vector<float> &array);
+int opesci_read_simple_binary_ptr(const char *filename, float *array, int size);
 int opesci_read_souces(const char *xyz_filename, const char *xsrc_filename, const char *ysrc_filename, const char *zsrc_filename,
 		       std::vector<float> &xyz_array, std::vector<float> &xsrc_array, std::vector<float> &ysrc_array, std::vector<float> &zsrc_array);
 int opesci_read_receivers(const char *filename, std::vector<float> &array);
